front: split scenario simulation report sheet
