+++ front/src/modules/simulationResult/components/SimulationResultExport/__tests__/utils.spec.ts
