+++ front/src/modules/simulationResult/components/SimulationResultExport/exportTrainCSV.ts
