+++ front/src/modules/simulationResult/components/SimulationResultExport/types.ts
