+++ front/src/modules/simulationResult/components/SimulationResultExport/utils.ts
